Removed unused code, refactored the directory
